backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
httpx
apscheduler
pandas
requests
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
xmltodict
google-generativeai
